requirements.txt
streamlit
numpy
pandas
Pillow
tensorflow==2.7.0